ResponseTrait -> HandlerTrait and ResponseTrait
