+++ src/App/src/Handler/ResponseTrait.php
@@ -4,78 +4,16 @@ declare(strict_types=1);
 
 namespace Api\App\Handler;
 
-use Api\App\Exception\BadRequestException;
-use Api\App\Exception\ConflictException;
-use Api\App\Exception\ExpiredException;
-use Api\App\Exception\ForbiddenException;
-use Api\App\Exception\MethodNotAllowedException;
-use Api\App\Exception\NotFoundException;
-use Api\App\Exception\UnauthorizedException;
-use Dot\Mail\Exception\MailException;
-use Exception;
-use Fig\Http\Message\RequestMethodInterface;
 use Fig\Http\Message\StatusCodeInterface;
 use Laminas\Diactoros\Response\EmptyResponse;
 use Laminas\Diactoros\Response\JsonResponse;
-use Laminas\Diactoros\Response\RedirectResponse;
-use Mezzio\Hal\HalResponseFactory;
-use Mezzio\Hal\Metadata\MetadataMap;
-use Mezzio\Hal\Metadata\RouteBasedCollectionMetadata;
-use Mezzio\Hal\ResourceGenerator;
-use Mezzio\Hal\ResourceGenerator\Exception\OutOfBoundsException;
-use Mezzio\Router\RouteResult;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use RuntimeException;
 
-use function array_key_exists;
 use function is_array;
-use function method_exists;
-use function sprintf;
-use function strtolower;
 
 trait ResponseTrait
 {
-    protected HalResponseFactory $responseFactory;
-    protected ResourceGenerator $resourceGenerator;
-
-    /**
-     * @throws Exception
-     */
-    public function handle(ServerRequestInterface $request): ResponseInterface
-    {
-        try {
-            $method = strtolower($request->getMethod());
-            if ($this->isGetCollectionRequest($request, $this->config)) {
-                $method = 'getCollection';
-            }
-
-            if (! method_exists($this, $method)) {
-                throw new MethodNotAllowedException(
-                    sprintf('Method %s is not implemented for the requested resource.', $method)
-                );
-            }
-
-            return $this->$method($request);
-        } catch (ConflictException $exception) {
-            return $this->errorResponse($exception->getMessage(), StatusCodeInterface::STATUS_CONFLICT);
-        } catch (ForbiddenException $exception) {
-            return $this->errorResponse($exception->getMessage(), StatusCodeInterface::STATUS_FORBIDDEN);
-        } catch (ExpiredException $exception) {
-            return $this->errorResponse($exception->getMessage(), StatusCodeInterface::STATUS_GONE);
-        } catch (OutOfBoundsException | NotFoundException $exception) {
-            return $this->errorResponse($exception->getMessage(), StatusCodeInterface::STATUS_NOT_FOUND);
-        } catch (UnauthorizedException $exception) {
-            return $this->errorResponse($exception->getMessage(), StatusCodeInterface::STATUS_UNAUTHORIZED);
-        } catch (MethodNotAllowedException $exception) {
-            return $this->errorResponse($exception->getMessage(), StatusCodeInterface::STATUS_METHOD_NOT_ALLOWED);
-        } catch (BadRequestException $exception) {
-            return $this->errorResponse($exception->getMessages(), StatusCodeInterface::STATUS_BAD_REQUEST);
-        } catch (MailException | RuntimeException | Exception $exception) {
-            return $this->errorResponse($exception->getMessage(), StatusCodeInterface::STATUS_INTERNAL_SERVER_ERROR);
-        }
-    }
-
     public function createResponse(ServerRequestInterface $request, mixed $instance): ResponseInterface
     {
         return $this->responseFactory->createResponse(
@@ -98,7 +36,7 @@ trait ResponseTrait
 
     public function errorResponse(
         array|string $messages = [],
-        int $status = StatusCodeInterface::STATUS_BAD_REQUEST
+        int $status = StatusCodeInterface::STATUS_INTERNAL_SERVER_ERROR
     ): ResponseInterface {
         return $this->jsonResponse([
             'error' => [
@@ -118,11 +56,6 @@ trait ResponseTrait
         ], $status);
     }
 
-    public function redirectResponse(string $location): ResponseInterface
-    {
-        return new RedirectResponse($location);
-    }
-
     public function jsonResponse(
         array|string $messages = [],
         int $status = StatusCodeInterface::STATUS_OK
@@ -130,35 +63,13 @@ trait ResponseTrait
         return new JsonResponse($messages, $status);
     }
 
-    /**
-     * @throws RuntimeException
-     */
-    private function isGetCollectionRequest(ServerRequestInterface $request, array $config): bool
+    public function notAcceptedResponse(string $message): ResponseInterface
     {
-        if ($request->getMethod() !== RequestMethodInterface::METHOD_GET) {
-            return false;
-        }
+        return $this->errorResponse($message, StatusCodeInterface::STATUS_NOT_ACCEPTABLE);
+    }
 
-        if (! array_key_exists(MetadataMap::class, $config)) {
-            throw new RuntimeException(
-                sprintf('Unable to load %s from config.', MetadataMap::class)
-            );
-        }
-
-        /** @var RouteResult $routeResult */
-        $routeResult = $request->getAttribute(RouteResult::class);
-        $routeName   = $routeResult->getMatchedRouteName();
-
-        $halConfig = null;
-        foreach ($config[MetadataMap::class] as $cfg) {
-            if ($cfg['route'] === $routeName) {
-                $halConfig = $cfg;
-                break;
-            }
-        }
-
-        return is_array($halConfig)
-            && array_key_exists('__class__', $halConfig)
-            && $halConfig['__class__'] === RouteBasedCollectionMetadata::class;
+    public function unsupportedMediaTypeResponse(string $message): ResponseInterface
+    {
+        return $this->errorResponse($message, StatusCodeInterface::STATUS_UNSUPPORTED_MEDIA_TYPE);
     }
 }
